Eradicate customlog.
This may well break stuff. The only way to be sure is to try it and see
if, well, stuff appears broken. Theoretically, this should clean up a
trouble-spot, as the customlog was a source of mess; the "uid" arg to
all messages may need to be dealt with, though, if I've missed any.