Remove the .source weakref which doesn't appear to be used.
TODO: Check if everything's working. If stuff starts breaking, revert this,
and then look into the pickling problem (saving an analysis to the db can't
work when there's an unpicklable object in the analysis, so the source link
is breaking that). It may be simplest to just remove the source link, build
the pickle, and then replace the source link; otherwise, actually implement
pickle protocol and specify what attributes are important to us.