renderer: Use time.perf_counter() instead of time.time()
If the clock sync issues are caused by time.time() not incrementing by 1.0
every second, it's possible that this will fix it. No guarantee though. So
as well as using perf_counter, this will repeatedly log the offset; if the
offset ever changes by a significant figure, you have clock drift.

May affect #237.